+++ savage_core/src/lib.rs
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: AGPL-3.0-or-later
 // Copyright (C) 2021-2022  Philipp Emanuel Weidmann <pew@worldwidemann.com>
 
-mod evaluate;
+pub mod evaluate;
 pub mod expression;
 pub mod functions;
 pub mod helpers;